pipeline/sagemaker/requirements.txt
jsonschema
numpy
pillow
pyarrow
tqdm
